+++ js/fullsize-picture.js
@@ -1,52 +1,21 @@
-import { isEscapeKey } from './itil.js';
-import { photos } from './gallery.js';
+import { getPhotos } from './script.js';
+import { renderThumbnails, template } from './thumbnail.js';
 
-//будем описывать логику показа и скрытия полноразмерного изображения
-const bigPictureModalElement = document.querySelector('.big-picture');
-const pictureModalOpenElements = Array.from(document.querySelectorAll('.picture'));
-const pictureModalCloseElement = document.querySelector('#picture-cancel');
-
-const onDocumentKeydown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    closePictureModal();
-  }
-};
-
-const onCloseIconClick = () => {
-  closePictureModal();
-};
-
-function openPictureModal() {
-  bigPictureModalElement.classList.remove('hidden');
-  bigPictureModalElement.querySelector('.social__comment-count').classList.add('hidden');
-  bigPictureModalElement.querySelector('.comments-loader').classList.add('hidden');
-  document.querySelector('body').classList.add('modal-open');
-
-  document.addEventListener('keydown', onDocumentKeydown);
-
-  pictureModalCloseElement.addEventListener('click', onCloseIconClick);
-}
-
-function closePictureModal() {
-  bigPictureModalElement.classList.add('hidden');
-  document.querySelector('body').classList.remove('modal-open');
-
-  document.removeEventListener('keydown', onDocumentKeydown);
-
-  pictureModalCloseElement.removeEventListener('click', onCloseIconClick);
-}
+export const photos = getPhotos();
+export const bigPictureModalElement = document.querySelector('.big-picture');
+renderThumbnails(photos);
 
 const commentsList = bigPictureModalElement.querySelector('.social__comments');
 
 const clearCommentsList = () => {
-  while (commentsList.firstChild) {
-    commentsList.removeChild(commentsList.firstChild);
+  if (commentsList.children) {
+    while (commentsList.firstChild) {
+      commentsList.removeChild(commentsList.firstChild);
+    }
   }
 };
 
 const renderCommentsList = (index) => {
-  const template = bigPictureModalElement.querySelector('ul').querySelector('.social__comment').cloneNode(true);
   const comments = photos[index].comments;
   const fragment = document.createDocumentFragment();
 
@@ -62,7 +31,7 @@ const renderCommentsList = (index) => {
   commentsList.appendChild(fragment);
 };
 
-const createFullsizePicture = (picture, index) => {
+export const createFullsizePicture = (picture, index) => {
   bigPictureModalElement.querySelector('.big-picture__img').querySelector('img').src = picture.querySelector('.picture__img').src;
   bigPictureModalElement.querySelector('.social').querySelector('span').textContent = picture.querySelector('.picture__likes').textContent;
   bigPictureModalElement.querySelector('.social__comment-count').querySelector('span').textContent = picture.querySelector('.picture__comments').textContent;
@@ -71,10 +40,3 @@ const createFullsizePicture = (picture, index) => {
   renderCommentsList(index);
 };
 
-pictureModalOpenElements.forEach((pictureModalOpenElement, index) => {
-  pictureModalOpenElement.addEventListener('click', () => {
-    createFullsizePicture(pictureModalOpenElement, index);
-
-    openPictureModal();
-  });
-});
