+++ js/itil.js
@@ -1,5 +1,3 @@
-
-
 const getRandomInteger = (a, b) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
