+++ js/thumbnail.js
@@ -19,4 +19,6 @@ const renderThumbnails = (photos) => {
   thumbnailsList.appendChild(fragment);
 };
 
+export const template = document.querySelector('.social__comment');
+
 export { renderThumbnails };
